Rewrite join
Due to a bug in the ByteStrings library, Data.ByteString.Lazy.append is
strict in the head of the second argument. This forces the next chunk, which
might not be available if you read from a handle, etc. This patch makes join
more lazy.

darcs-hash:20070303070850-6bdf4-148e1f25b2ae1825fe473524b4f56acb99254878.gz