yay another speedup + benchmarking against @kolmodin's cps branch
The code in blaze-binary is at 1.5x faster for decoding and 2x faster for
encoding than the newest `binary` code. For nested datastructures, e.g., a
list of lists, the speedup becomes more like a factor 4.

I suspect that unknown calls are cheaper for my representation, as the parsing
state is hoisted out into the driver. Moreover, we now use a different list
encoding: we encode lists size-prefixed and in reverse order. This simplifies
decoding and does not really slow-down encoding, as reversing and measuring
the length can be done at once. It halves the number of writes and reads of
primitive values, which compensates for traversing the list twice.